requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.8
idna==2.6
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
